backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
opencv-python==4.9.0.80
numpy==1.26.3
Pillow==10.2.0
mediapipe==0.10.14
rembg==2.0.56
